feat: compare card s owner & user while delete card

+++ controllers/cards.js
@@ -2,7 +2,8 @@ const Card = require('../models/card');
 
 const BadRequestError = require('../utils/errors/BadRequestError');
 const NotFoundError = require('../utils/errors/NotFoundError');
-const UnauthorizedError = require('../utils/errors/UnauthorizedError');
+// const UnauthorizedError = require('../utils/errors/UnauthorizedError');
+const ForbiddenError = require('../utils/errors/ForbiddenError');
 
 const { checkIdValidity } = require('../utils/checkIdValidity');
 const { setErrorType } = require('../utils/setErrorType');
@@ -35,14 +36,15 @@ const deleteCard = (req, res, next) => {
   if (!checkIdValidity(req.params.cardId)) {
     throw new BadRequestError();
   }
+  if (req.body.owner === req.user._id) {
+    throw new ForbiddenError();
+  }
+
   Card.findByIdAndRemove(req.params.cardId)
     .then((card) => {
       if (!card) {
         throw new NotFoundError();
       }
-      // if (req.params.cardId !== req.user._id) {
-      //   throw new UnauthorizedError();
-      // }
 
       res.send(card);
     })
